# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Foldways API",
    "description": "Protein engineering tools running on Modal.",
    "version": "0.3.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Get Health",
        "operationId": "get_health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/services": {
      "get": {
        "summary": "List Services",
        "description": "List the available services and how to call each.",
        "operationId": "list_services_services_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceInfo"
                  },
                  "type": "array",
                  "title": "Response List Services Services Get"
                }
              }
            }
          }
        }
      }
    },
    "/services/{name}": {
      "get": {
        "summary": "Get Service",
        "description": "Describe a service: its description and the params to submit.",
        "operationId": "get_service_services__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "summary": "List Jobs",
        "description": "List all jobs and their current status.",
        "operationId": "list_jobs_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobStatus"
                  },
                  "type": "array",
                  "title": "Response List Jobs Jobs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Submit Job",
        "description": "Submit a single job naming a service and its params.\n\nThe job runs as an independent Modal call with no parent batch. Returns a\nserver-generated id used for status, download, and deletion.",
        "operationId": "submit_job_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "description": "Get a single job's status.",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Stop And Delete Job",
        "description": "Stop a job if running and delete it along with its stored data.",
        "operationId": "stop_and_delete_job_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/download": {
      "get": {
        "summary": "Download Job",
        "description": "Download a completed job's output directory as a zip archive.\n\nErrors with 409 if the output isn't ready yet (job pending or failed).",
        "operationId": "download_job_jobs__job_id__download_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/stop": {
      "post": {
        "summary": "Stop Job",
        "description": "Stop a running job without deleting it or its data.",
        "operationId": "stop_job_jobs__job_id__stop_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/batches": {
      "post": {
        "summary": "Submit Batch",
        "description": "Submit a sweep: run one service over many param sets.\n\nEach param set becomes an independent Modal call under a shared batch id,\nnamed \"{name}-{i}\" when a batch name is given. Returns the batch id plus a\nserver-generated id per job; use either for status and deletion, and each\njob id with GET /jobs/{job_id}/download to fetch that job's output.",
        "operationId": "submit_batch_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/batches/{batch_id}": {
      "get": {
        "summary": "Get Batch",
        "description": "Get a batch's aggregate status and the status of each job in it.",
        "operationId": "get_batch_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Batch",
        "description": "Stop a batch and delete every job in it along with their stored data.",
        "operationId": "delete_batch_batches__batch_id__delete",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/batches/{batch_id}/stop": {
      "post": {
        "summary": "Stop Batch",
        "description": "Stop all running jobs in a batch without deleting them.",
        "operationId": "stop_batch_batches__batch_id__stop_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BatchRequest": {
        "properties": {
          "service": {
            "type": "string",
            "title": "Service"
          },
          "params": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Params"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "mock": {
            "type": "boolean",
            "title": "Mock",
            "default": false
          }
        },
        "type": "object",
        "required": ["service", "params"],
        "title": "BatchRequest"
      },
      "BatchState": {
        "type": "string",
        "enum": ["in_progress", "completed", "completed_with_failures"],
        "title": "BatchState",
        "description": "The status values a batch can report, aggregated from its jobs."
      },
      "BatchStatus": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "status": {
            "$ref": "#/components/schemas/BatchState"
          },
          "counts": {
            "additionalProperties": true,
            "type": "object",
            "title": "Counts"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobStatus"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": ["id", "created_at", "status", "counts", "jobs"],
        "title": "BatchStatus"
      },
      "BatchSubmitResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobStatus"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": ["batch_id", "jobs"],
        "title": "BatchSubmitResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobRequest": {
        "properties": {
          "service": {
            "type": "string",
            "title": "Service"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "mock": {
            "type": "boolean",
            "title": "Mock",
            "default": false
          }
        },
        "type": "object",
        "required": ["service", "params"],
        "title": "JobRequest"
      },
      "JobState": {
        "type": "string",
        "enum": [
          "pending",
          "complete",
          "failed",
          "init_failed",
          "stopped",
          "timed_out",
          "unknown"
        ],
        "title": "JobState",
        "description": "The status values a single job can report."
      },
      "JobStatus": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "service": {
            "type": "string",
            "title": "Service"
          },
          "status": {
            "$ref": "#/components/schemas/JobState"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          }
        },
        "type": "object",
        "required": ["id", "service", "status", "created_at"],
        "title": "JobStatus"
      },
      "ServiceInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "params_schema": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params Schema"
          }
        },
        "type": "object",
        "required": ["name", "description", "params_schema"],
        "title": "ServiceInfo"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}